Add search and pagination to user index
The "mazer" layout is no longer used as we have moved all the code on
application.html.erb.

We are not using ApexCharts currently so I removed the lines that were
importing scripts for that.

I tried writing system tests to test the search and pagination but I
couldn't find how to make Capybara see it. It worked when I tested it
by hand.