Add search functionality to users
There used to be a search functionality in the users index page that was
using Mazer's datatable stylesheet (PR#52), but it was then removed in
PR#86 as we discussed doing it our way instead.

This commit enables searching Users by using a similar logic to what
Dmitry did for searching Topics. For the JS controller, I re-used what
we had in topics_controller.js but did not replace topics_controller.js
with the more generic "search_controller.js" after discussion with
Dmitry during PR review. This is because the chooseProvider function is
specific to the topics controller and replacing it with a generic name
such as chooseOption may be too generic if we then need to add other
choosing actions on Topics.