Remove button to upload files
We had originally added a button like in the article we followed, but
felt that it would be more user-friendly to automatically upload the
file once it is selected.

For some reason the test for editing a topic in upload_management_spec.rb
was failing on my machine because it either couldn't see the Topic or
was on the Dashboard instead of being on the "Topics" page. I'm not sure
why this was an issue since it was working on Dalyl's computer and the CI,
but moving "click_link("Topics")" to inside the "it" blocks solved the
issue.