Display errors when topic creation fails
I changed the feature test file to a system test file
because for some reason the assertions were passing even when
I could see they shouldn't when testing manually.

It seems like we are hiding the provider dropdown from Contributors
but Topic creation fails when the provider is not selected, I wasn't
sure what the requirements were so I just wrote the tests for Contributors
and left the one that doesn't pass as pending for now.