Don't attach documents unless confirmed by user
This code was attaching documents to a pre-existing Topic as soon as
it was uploaded, but I think that would be confusing if the user uploaded
a document and then changed their mind and clicked on "Cancel". It should
only be attached once the user has confirmed that they want to attach it
by clicking on "Update Topic".