Ensure document-related errors are displayed
I saw the DocumentValidator is not being used so I removed it.

I wasn't sure what the best way was to validate the blobs, if we should
do it right away when creating them in the UploadsController but that
made it more complicated to show the errors on the form.

I added PDF as a valid content_type following Danny's comment.