Fix Topic scope issue on Topics#index
For Admins as for Contributors, we want to only display the topics that
belong to the first provider they are entitled to see, as this provider's
name is displayed by default on the index page. Users are then able to
see topics from other providers they are entitled to check by selecting
them from the dropdown menu.

The provider_id param is not part of the search form, it is part of the
provider form in the choose_provider partial that allows us to set the
current_provider_id cookie in SettingsController, so it wasn't being
passed. We can ensure we see only relevant topics by redefining #scope,
which will change according to the presence of current_provider_id cookie,
instead of having a wider scope and then trying to check the params.