Remove unused User#topics method
I think this method is currently unnecessary, as we are only showing
topics scoped by provider. But if we wanted something like this, I would
like to suggest using an association:

has_many :topics, through: :providers"

or, if we want to keep this as a method, change:

Topic.where(provider_id: providers.pluck(:id))

to:

Topic.where(provider_id: providers)

as it would allow us to get the topics in 1 query instead of having 1
query for plucking the providers ids and then another for getting the topics.