Move provider association management to user form
Until now, we only allowed adding a user to a provider on the provider
create or edit form, but the other users belonging to that provider were
deselected when choosing a new user to add without pressing the multiselect
key, so there was a risk of inadvertently removing users.

This commit moves the provider association management to the user form
as this is probably more user-friendly, as discussed with Danny on Zoom
and mentioned in issue 161.

https://github.com/rubyforgood/skillrx/issues/161

I created a selectable-controller.js extracting just what I needed from
the select-tags-controller.js because I was concerned it would be
confusing re-using the select-tags-controller as it includes a lot of
logic that is only valid for Tags. The fact that these are called
"Bootstrap tags" but it matches a model name also made it a bit confusing.

I also removed some duplicate ids as there were errors in the console
concerning this.