Fix issue with adding new cognates to Tags
This resolves issue #177. Adding a new cognate to a Tag was failing
silently and showing the form without displaying any errors. I saw in the
logs that the problem happened because we tried to create a tag_cognate
again for the Cognates that were already associated with the Tag, so I
changed #create to #find_or_create_by.