Fix database setup username configuration
After PR #199 was merged, bin/dev stopped working for me and I started
being prompted for a password when typing rails db. Danny and I looked
into it and noticed that if DATABASE_USER was not set in my ENV file,
the user did not fallback to "skillrx" unless I changed the "user" key to
"username" in database.yml.

There was also a mismatch between DATABASE_USERNAME in database.yml and
DATABASE_USER in .env.example and docker-compose.dev.yml, so I changed
DATABASE_USERNAME to DATABASE_USER.