Fix database setup configuration
After PR #199 was merged, I started seeing this error when running bin/dev:
"PG::ConnectionBad (could not translate host name "db" to address:
Temporary failure in name resolution)". I saw the same error when running
rails db.

Danny and I looked into it and saw we needed to re-instate DATABASE_URL
and QUEUE_DATABASE_URL in order to avoid getting this error

I was also prompted for a password and noticed that if DATABASE_USER
was not set in my ENV file, the user did not fallback to "skillrx" unless
I changed the "user" key to "username" in database.yml.