Add search for Tag index
Currently, if we have a mix of tags starting with a capital letter and
others starting with a lowercase letter, the lower-cased tags will be
displayed at the end. I considered changing the sorting, but I think
it would probably be better to fix the tags' names to make them consistent
and ensure consistent capitalisation when we save the tags.