Ensure cognates get added to/removed from topics
As discussed with Danny on Slack, if users associate a Tag to a Topic,
that Tag's cognates/reverse cognates must also be added to the Topic.
Likewise, if a Tag is removed from a Topic, its cognates/reverse cognates
should be removed as well.

A Tag can have cognates in English and Spanish, so we should only add
the cognates that match the Topic's language. For example, there is
only 1 tag called "malaria", but it's the same word in English and
Spanish so it can be tagged on Topics of either language, but its
cognates may be distinctively English or Spanish.