Synchronize Topics when a Tag has a new cognate
Currently, when users tag a Topic, we automatically tag that Topic with
that Tag's cognates.

This commit introduces a job to make sure that we also tag Topic with
any relevant cognates when an Admin adds a cognate to a Tag.

I haven't done anything for when an Admin removes a cognate from a Tag
as I think the removed cognate might still be relevant to the Topic,
even if it is no longer as a cognate.