Change CSV row separator to not use CRLF
Manoj said:

"It is important that the .csv file does not have any CRLF (\r) characters
in it, as it breaks the import of the .csv into local database on the
edge device. It seems that windows environment adds \r, which is not
visible in the .csv file (compared to \n used on linux)"

According to the CSV documentation, by default, the row separator is
"\r\n" (CRLF). So I changed it to just "\n" (LF) in the hope it would
solve the issue.

https://ruby.github.io/csv/doc/csv/recipes/generating_rdoc.html#label-Row+Separator