Set AuthorID to 0 in TopicAuthor.csv
Here is the explanation from Danny:

"We actually don't want to deliver real author data here. The downstream
application has a concept of authors, with an authors table and author
association with any topic. Those authors are not users in the system;
the editors add authors as necessary.

We eliminated that early in the requirements process but they did not
adjust their downstream app's requirements, so it is expecting an authors
csv with one entry for each topic and an author ID. They have asked us
to deliver that file with 0 for each of the author IDs. So line 20 can
just return a zero."