Fix spec/system/topics/search_spec.rb tests
The test checking that no dropdown menu is present when a user only has
1 provider was previously passing erroneously: it was checking that no
provider select menu was present before the Topics page was loaded, so it
only confirmed that that dropdown menu was not on the "Dashbaord" page.

I had to tweak some tests because:

- It is also no longer possible for a contributor to have no providers.
- we now automatically associate 1 provider to the user in the factory,
so the set up for "with 1 associated provider" was incorrect.