Move access-related tests to authorizations_spec
We already had tests to check that only Admins could access Mission Control.
But now that the link to "Jobs" is in the sidebar together with other
pages for which we test who has access in authorizations_spec, I think
it's clearer for us to have all these tests in the same file, so I moved
the tests from jobs_spec.rb to authorizations_spec.rb and added tests for
when the user is not authenticated.