Remove language context from Taggable concern
Currently, Tags are associated to a Topic via a Tagging with a context
set to the language code of the Topic. This commit makes changes so we
ignore this context and get all associated tags regardless of context.
Once this is merged and works with the current data, we'll fix the
existing Taggings' "context" column to be "tags" so we can use
acts_as_taggable methods such as "tags".