Remove button to add new Tags from Tag index
Removing the ability to add new Tags from the Tag index was a conscious
decision we took because we didn't want Tags that were not used on any
Topics, and at the time it would not have been possible to determine
for which language a Tag was meant. Now that we display all Tags
regardless of a Topic's language, the language consideration would no
longer be an issue, but the stakeholders haven't told us they needed
this feature back.

See the PR where we removed the ability to add Tags from the Tag index:
https://github.com/rubyforgood/skillrx/pull/233/commits/8f7703034dd7de1d2faa3cf359b1515a76938aa2