Change color buttons on Topic show page
Red is generally displayed for destructive actions, but here we were
displaying "Edit" in red and "Archive" in orange. Since we are using
orange on the Edit forms elsewhere, I've used orange for the Edit button,
gray for Archive, and red for Delete. But orange is also often used to
indicate a mildly dangerous action, so I wonder if it should be changed
to something less alarming, like blue, purple or green. But editing a
Topic can indeed be slightly dangerous if the user makes a mistake, so
it could also be argued that orange is appropriate.